# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample online-shop server",
        "title": "Online shop API Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/orders": {
            "post": {
                "description": "Создает заказ с указанным списком продуктов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Создать новый заказ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Токен",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Информация о заказе",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.OrderData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешное создание заказа",
                        "schema": {
                            "$ref": "#/definitions/types.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос (например, пустой список продуктов)",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера (например, не удалось создать заказ или добавить продукт)",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "description": "Извлекает детали конкретного заказа, включая продукты, связанные с ним.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Получить детали заказа по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Токен",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID заказа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Детали заказа",
                        "schema": {
                            "$ref": "#/definitions/types.Order"
                        }
                    },
                    "400": {
                        "description": "Неверный ID заказа",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Заказ не найден",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "types.Category": {
            "type": "object",
            "properties": {
                "categoryID": {
                    "description": "Primary key",
                    "type": "integer"
                },
                "description": {
                    "description": "Description of the category",
                    "type": "string"
                },
                "name": {
                    "description": "Category name",
                    "type": "string"
                }
            }
        },
        "types.Characteristic": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "Name of category",
                    "type": "string"
                },
                "characteristicID": {
                    "description": "Primary key",
                    "type": "integer"
                },
                "name": {
                    "description": "Name of the characteristic",
                    "type": "string"
                }
            }
        },
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "types.Manufacturer": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "Description of the manufacturer",
                    "type": "string"
                },
                "manufacturerID": {
                    "description": "Primary key",
                    "type": "integer"
                },
                "name": {
                    "description": "Manufacturer name",
                    "type": "string"
                }
            }
        },
        "types.Order": {
            "type": "object",
            "properties": {
                "order_id": {
                    "description": "Уникальный идентификатор заказа",
                    "type": "integer"
                },
                "products": {
                    "description": "Список продуктов в заказе",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ProductOrder"
                    }
                },
                "status": {
                    "description": "Статус заказа",
                    "type": "string"
                },
                "time_stamp": {
                    "description": "Временная метка создания заказа",
                    "type": "string"
                }
            }
        },
        "types.OrderData": {
            "type": "object",
            "properties": {
                "products": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "count": {
                                "type": "integer"
                            },
                            "product_id": {
                                "type": "integer"
                            }
                        }
                    }
                }
            }
        },
        "types.Product": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ProductCategory"
                    }
                },
                "characteristic": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ProductCharacteristic"
                    }
                },
                "cost": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ProductCost"
                    }
                },
                "description": {
                    "description": "Product description",
                    "type": "string"
                },
                "imageLink": {
                    "description": "URL to the product image",
                    "type": "string"
                },
                "manufacturer": {
                    "description": "Associations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.Manufacturer"
                        }
                    ]
                },
                "manufacturer_id": {
                    "description": "Foreign key to Manufacturer",
                    "type": "integer"
                },
                "name": {
                    "description": "Product name",
                    "type": "string"
                },
                "product_id": {
                    "description": "Primary key",
                    "type": "integer"
                },
                "status": {
                    "description": "Status",
                    "type": "string"
                }
            }
        },
        "types.ProductCategory": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/types.Category"
                }
            }
        },
        "types.ProductCharacteristic": {
            "type": "object",
            "properties": {
                "characteristic": {
                    "$ref": "#/definitions/types.Characteristic"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "types.ProductCost": {
            "type": "object",
            "properties": {
                "costId": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                },
                "productId": {
                    "type": "integer"
                },
                "startTimeStamp": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
...
        "types.ProductOrder": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "Количество продукта",
                    "type": "integer"
                },
                "order_id": {
                    "description": "ID заказа (внешний ключ)",
                    "type": "integer"
                },
                "product": {
                    "$ref": "#/definitions/types.Product"
                },
                "product_id": {
                    "description": "ID продукта (внешний ключ)",
                    "type": "integer"
                }
            }
        },
        "types.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
    },
...
}